% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_info.R
\name{extract_config_framework}
\alias{extract_config_framework}
\title{extract_config_framework - the configuration framework for GeneCodeR}
\usage{
extract_config_framework(verbose = F)
}
\arguments{
\item{verbose}{- set to TRUE to print out the config parameters, to be editted}
}
\description{
extract_config_framework - the configuration framework for GeneCodeR
}
