% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_info.R
\name{extract_meta_framework}
\alias{extract_meta_framework}
\title{extract_meta_framework - the meta information framework for GeneCodeR}
\usage{
extract_meta_framework(verbose = F)
}
\arguments{
\item{verbose}{- set to TRUE to print out the meta information, to be editted}
}
\description{
extract_meta_framework - the meta information framework for GeneCodeR
}
