% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_info.R
\name{extract_path_framework}
\alias{extract_path_framework}
\title{extract_path_framework - the path list framework for GeneCodeR}
\usage{
extract_path_framework(verbose = F)
}
\arguments{
\item{verbose}{- set to TRUE to print out the path list, to be editted}
}
\description{
extract_path_framework - the path list framework for GeneCodeR
}
