% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genecoder.R
\name{genecoder}
\alias{genecoder}
\title{GeneCodeR - Perturbing imaging tissue by altering spatial gene codes}
\usage{
genecoder(model, x, config)
}
\description{
GeneCodeR - Perturbing imaging tissue by altering spatial gene codes
}
