% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genecoder.R
\name{genecoder.transform}
\alias{genecoder.transform}
\title{genecoder.transform - transform between modalities}
\usage{
genecoder.transform(model, x, config.transform)
}
\arguments{
\item{model}{- the learned model via Generative Encoding}

\item{x}{- the dataset to be transformed}

\item{config.transform}{- the transform list within the GeneCodeR configuration parameters to define the modality coming from and going to}
}
\description{
genecoder.transform - transform between modalities
}
