% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genecoder.R
\name{learn_model}
\alias{learn_model}
\title{learn_model - learns the Generative Encoder model}
\usage{
learn_model(
  data_list,
  config = gcode::extract_config(F),
  transfer = gcode::extract_transfer_framework(F),
  recover = gcode::extract_recovery_framework(F),
  join,
  reference
)
}
\arguments{
\item{data_list}{List of data matrices of varying dimensionality. Attempts to find similarities among all datasets with a core structure.}

\item{config}{Configuration parameters (required, default provided)}

\item{transfer}{Transferring pre-trained model parameters (not required)}

\item{recover}{Important information used for prediction or imputation (not required)}

\item{join}{Join parameters that share the same axes to be similar (not required)

#' @export}
}
\description{
learn_model - learns the Generative Encoder model
}
