% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genecoder.R
\name{prepare_gex}
\alias{prepare_gex}
\title{prepare_gex - prepares gene expression data}
\usage{
prepare_gex(file_path_list, meta_info_list, config)
}
\arguments{
\item{file_path_list}{- a list of files containing the path to all files including the gene expression path}

\item{meta_info_list}{- a list of meta information that helps read in files and extracts relevant information}

\item{config}{- the main configuration parameters to extract the image pixels for each corresponding spot}
}
\description{
prepare_gex - prepares gene expression data
}
