% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genecoder.R
\name{prepare_spot}
\alias{prepare_spot}
\title{prepare_spot - prepares image information per spot}
\usage{
prepare_spot(file_path_list, meta_info_list, config, gex_data)
}
\arguments{
\item{file_path_list}{- a list of files containing the path to all files including the image path}

\item{meta_info_list}{- a list of meta information that helps read in files and extracts relevant information}

\item{config}{- the main configuration parameters to extract the image pixels for each corresponding spot}
}
\description{
prepare_spot - prepares image information per spot
}
