% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genecoder.R
\name{genecoder}
\alias{genecoder}
\title{GeneCodeR - Perturbing imaging tissue by altering spatial gene codes}
\usage{
genecoder(model, x, model_type, config)
}
\arguments{
\item{model}{- the learned model via Generative Encoding}

\item{x}{- the dataset to be transformed}

\item{config}{- the GeneCodeR configuration parameters to define the modality coming from and going to}
}
\description{
GeneCodeR - Perturbing imaging tissue by altering spatial gene codes
}
